% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimisation.R
\name{RunClusterOptimisation}
\alias{RunClusterOptimisation}
\title{Cluster optimisation function.}
\usage{
RunClusterOptimisation(propRuns, intLength)
}
\arguments{
\item{propRuns}{proportion of total simulations to sample during optimisation}

\item{intLength}{integer defining number of individual intervention permutations to simulate}
}
\description{
This function controls the entire optimisation procedure in CascadeDashboard, and is a port of
}

